Use numeric value for alpha (#22)
What does this Pull Request accomplish?
I tried running the notebook on a SystemLink install and ran into an error "TypeError: alpha must be a float or None".

This PR fixes the error by using a float instead of a string for the alpha parameter to matplotlib.pyplot.scatter. matplotlib docs

Why should this Pull Request be merged?
The notebook should work out of the box.

What testing has been done?
Ran the notebook with my changes and verified that no errors occur.